Bundle discount

+++ java/src/main/java/dojo/supermarket/model/Discount.java
@@ -1,12 +1,14 @@
 package dojo.supermarket.model;
 
+import java.math.BigDecimal;
+
 public class Discount {
 
     private final String description;
-    private final double discountAmount;
+    private final BigDecimal discountAmount;
     private final Product product;
 
-    public Discount(Product product, String description, double discountAmount) {
+    public Discount(Product product, String description, BigDecimal discountAmount) {
         this.product = product;
         this.description = description;
         this.discountAmount = discountAmount;
@@ -16,7 +18,7 @@ public class Discount {
         return description;
     }
 
-    public double getDiscountAmount() {
+    public BigDecimal getDiscountAmount() {
         return discountAmount;
     }
